refactor Coordinator and CoordinatorPrepare
